move some tool functions
