+++ rdagent/log/ui/ds_trace.py
@@ -11,10 +11,9 @@ import streamlit as st
 from streamlit import session_state as state
 
 from rdagent.app.data_science.loop import DataScienceRDLoop
-from rdagent.log.mle_summary import extract_mle_json, is_valid_session
 from rdagent.log.storage import FileStorage
 from rdagent.log.ui.utils import load_times
-from rdagent.log.utils import extract_evoid, extract_loopid_func_name
+from rdagent.log.utils import extract_evoid, extract_loopid_func_name, is_valid_session, extract_mle_json
 from rdagent.utils import remove_ansi_codes
 from rdagent.utils.repo.diff import generate_diff_from_dict
