+++ rdagent/log/utils/__init__.py
@@ -1,8 +1,8 @@
 import inspect
 import re
 from typing import Optional, TypedDict, cast
-
-
+from pathlib import Path
+import json
 class LogColors:
     """
     ANSI color codes for use in console output.
@@ -77,6 +77,10 @@ def get_caller_info() -> CallerInfo:
     return info
 
 
+def is_valid_session(log_path: Path) -> bool:
+    return log_path.is_dir() and log_path.joinpath("__session__").exists()
+
+
 def extract_loopid_func_name(tag: str) -> tuple[str, str] | tuple[None, None]:
     """extract loop id and function name from the tag in Message"""
     match = re.search(r"Loop_(\d+)\.([^.]+)", tag)
@@ -87,3 +91,10 @@ def extract_evoid(tag: str) -> str | None:
     """extract evo id from the tag in Message"""
     match = re.search(r"\.evo_loop_(\d+)\.", tag)
     return cast(str, match.group(1)) if match else None
+
+
+def extract_mle_json(log_content: str) -> dict | None:
+    match = re.search(r"\{.*\}", log_content, re.DOTALL)
+    if match:
+        return json.loads(match.group(0))
+    return None
